Update unit tests and bazel files
