+++ pkg/kubelet/apis/BUILD
@@ -10,8 +10,20 @@ go_library(
     srcs = [
         "well_known_annotations.go",
         "well_known_labels.go",
-    ],
+    ] + select({
+        "@io_bazel_rules_go//go/platform:windows": [
+            "well_known_annotations_windows.go",
+        ],
+        "//conditions:default": [],
+    }),
     importpath = "k8s.io/kubernetes/pkg/kubelet/apis",
+    deps = select({
+        "@io_bazel_rules_go//go/platform:windows": [
+            "//pkg/features:go_default_library",
+            "//vendor/k8s.io/apiserver/pkg/util/feature:go_default_library",
+        ],
+        "//conditions:default": [],
+    }),
 )
 
 filegroup(
