+++ pkg/kubelet/kuberuntime/BUILD
@@ -22,7 +22,53 @@ go_library(
         "labels.go",
         "legacy.go",
         "security_context.go",
-    ],
+    ] + select({
+        "@io_bazel_rules_go//go/platform:android": [
+            "helpers_unsupported.go",
+            "kuberuntime_container_unsupported.go",
+        ],
+        "@io_bazel_rules_go//go/platform:darwin": [
+            "helpers_unsupported.go",
+            "kuberuntime_container_unsupported.go",
+        ],
+        "@io_bazel_rules_go//go/platform:dragonfly": [
+            "helpers_unsupported.go",
+            "kuberuntime_container_unsupported.go",
+        ],
+        "@io_bazel_rules_go//go/platform:freebsd": [
+            "helpers_unsupported.go",
+            "kuberuntime_container_unsupported.go",
+        ],
+        "@io_bazel_rules_go//go/platform:linux": [
+            "helpers_linux.go",
+            "kuberuntime_container_linux.go",
+        ],
+        "@io_bazel_rules_go//go/platform:nacl": [
+            "helpers_unsupported.go",
+            "kuberuntime_container_unsupported.go",
+        ],
+        "@io_bazel_rules_go//go/platform:netbsd": [
+            "helpers_unsupported.go",
+            "kuberuntime_container_unsupported.go",
+        ],
+        "@io_bazel_rules_go//go/platform:openbsd": [
+            "helpers_unsupported.go",
+            "kuberuntime_container_unsupported.go",
+        ],
+        "@io_bazel_rules_go//go/platform:plan9": [
+            "helpers_unsupported.go",
+            "kuberuntime_container_unsupported.go",
+        ],
+        "@io_bazel_rules_go//go/platform:solaris": [
+            "helpers_unsupported.go",
+            "kuberuntime_container_unsupported.go",
+        ],
+        "@io_bazel_rules_go//go/platform:windows": [
+            "helpers_windows.go",
+            "kuberuntime_container_windows.go",
+        ],
+        "//conditions:default": [],
+    }),
     importpath = "k8s.io/kubernetes/pkg/kubelet/kuberuntime",
     deps = [
         "//pkg/api/legacyscheme:go_default_library",
@@ -40,7 +86,6 @@ go_library(
         "//pkg/kubelet/lifecycle:go_default_library",
         "//pkg/kubelet/metrics:go_default_library",
         "//pkg/kubelet/prober/results:go_default_library",
-        "//pkg/kubelet/qos:go_default_library",
         "//pkg/kubelet/types:go_default_library",
         "//pkg/kubelet/util/cache:go_default_library",
         "//pkg/kubelet/util/format:go_default_library",
@@ -64,7 +109,16 @@ go_library(
         "//vendor/k8s.io/client-go/tools/record:go_default_library",
         "//vendor/k8s.io/client-go/tools/reference:go_default_library",
         "//vendor/k8s.io/client-go/util/flowcontrol:go_default_library",
-    ],
+    ] + select({
+        "@io_bazel_rules_go//go/platform:linux": [
+            "//pkg/kubelet/qos:go_default_library",
+        ],
+        "@io_bazel_rules_go//go/platform:windows": [
+            "//pkg/kubelet/apis:go_default_library",
+            "//vendor/github.com/docker/docker/pkg/sysinfo:go_default_library",
+        ],
+        "//conditions:default": [],
+    }),
 )
 
 go_test(
@@ -80,7 +134,12 @@ go_test(
         "labels_test.go",
         "legacy_test.go",
         "security_context_test.go",
-    ],
+    ] + select({
+        "@io_bazel_rules_go//go/platform:linux": [
+            "kuberuntime_container_linux_test.go",
+        ],
+        "//conditions:default": [],
+    }),
     embed = [":go_default_library"],
     deps = [
         "//pkg/credentialprovider:go_default_library",
