+++ pkg/kubelet/kuberuntime/kuberuntime_container_test.go
@@ -203,107 +203,6 @@ func TestToKubeContainerStatus(t *testing.T) {
 	}
 }
 
-func makeExpectedConfig(m *kubeGenericRuntimeManager, pod *v1.Pod, containerIndex int) *runtimeapi.ContainerConfig {
-	container := &pod.Spec.Containers[containerIndex]
-	podIP := ""
-	restartCount := 0
-	opts, _ := m.runtimeHelper.GenerateRunContainerOptions(pod, container, podIP)
-	containerLogsPath := buildContainerLogsPath(container.Name, restartCount)
-	restartCountUint32 := uint32(restartCount)
-	envs := make([]*runtimeapi.KeyValue, len(opts.Envs))
-
-	expectedConfig := &runtimeapi.ContainerConfig{
-		Metadata: &runtimeapi.ContainerMetadata{
-			Name:    container.Name,
-			Attempt: restartCountUint32,
-		},
-		Image:       &runtimeapi.ImageSpec{Image: container.Image},
-		Command:     container.Command,
-		Args:        []string(nil),
-		WorkingDir:  container.WorkingDir,
-		Labels:      newContainerLabels(container, pod, kubecontainer.ContainerTypeRegular),
-		Annotations: newContainerAnnotations(container, pod, restartCount, opts),
-		Devices:     makeDevices(opts),
-		Mounts:      m.makeMounts(opts, container),
-		LogPath:     containerLogsPath,
-		Stdin:       container.Stdin,
-		StdinOnce:   container.StdinOnce,
-		Tty:         container.TTY,
-		Linux:       m.generateLinuxContainerConfig(container, pod, new(int64), ""),
-		Envs:        envs,
-	}
-	return expectedConfig
-}
-
-func TestGenerateContainerConfig(t *testing.T) {
-	_, imageService, m, err := createTestRuntimeManager()
-	assert.NoError(t, err)
-
-	pod := &v1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			UID:       "12345678",
-			Name:      "bar",
-			Namespace: "new",
-		},
-		Spec: v1.PodSpec{
-			Containers: []v1.Container{
-				{
-					Name:            "foo",
-					Image:           "busybox",
-					ImagePullPolicy: v1.PullIfNotPresent,
-					Command:         []string{"testCommand"},
-					WorkingDir:      "testWorkingDir",
-				},
-			},
-		},
-	}
-
-	expectedConfig := makeExpectedConfig(m, pod, 0)
-	containerConfig, err := m.generateContainerConfig(&pod.Spec.Containers[0], pod, 0, "", pod.Spec.Containers[0].Image, kubecontainer.ContainerTypeRegular)
-	assert.NoError(t, err)
-	assert.Equal(t, expectedConfig, containerConfig, "generate container config for kubelet runtime v1.")
-
-	runAsUser := int64(0)
-	runAsNonRootTrue := true
-	podWithContainerSecurityContext := &v1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			UID:       "12345678",
-			Name:      "bar",
-			Namespace: "new",
-		},
-		Spec: v1.PodSpec{
-			Containers: []v1.Container{
-				{
-					Name:            "foo",
-					Image:           "busybox",
-					ImagePullPolicy: v1.PullIfNotPresent,
-					Command:         []string{"testCommand"},
-					WorkingDir:      "testWorkingDir",
-					SecurityContext: &v1.SecurityContext{
-						RunAsNonRoot: &runAsNonRootTrue,
-						RunAsUser:    &runAsUser,
-					},
-				},
-			},
-		},
-	}
-
-	_, err = m.generateContainerConfig(&podWithContainerSecurityContext.Spec.Containers[0], podWithContainerSecurityContext, 0, "", podWithContainerSecurityContext.Spec.Containers[0].Image, kubecontainer.ContainerTypeRegular)
-	assert.Error(t, err)
-
-	imageId, _ := imageService.PullImage(&runtimeapi.ImageSpec{Image: "busybox"}, nil)
-	image, _ := imageService.ImageStatus(&runtimeapi.ImageSpec{Image: imageId})
-
-	image.Uid = nil
-	image.Username = "test"
-
-	podWithContainerSecurityContext.Spec.Containers[0].SecurityContext.RunAsUser = nil
-	podWithContainerSecurityContext.Spec.Containers[0].SecurityContext.RunAsNonRoot = &runAsNonRootTrue
-
-	_, err = m.generateContainerConfig(&podWithContainerSecurityContext.Spec.Containers[0], podWithContainerSecurityContext, 0, "", podWithContainerSecurityContext.Spec.Containers[0].Image, kubecontainer.ContainerTypeRegular)
-	assert.Error(t, err, "RunAsNonRoot should fail for non-numeric username")
-}
-
 func TestLifeCycleHook(t *testing.T) {
 
 	// Setup
